[PIR #1553] Move Attention projection into codev-sdk/builder-helpers; add idle-waiting to the rollup
